feat : 렌더링 속도 변경 노력
