--- src/components/book/MakeBookButton.tsx
@@ -1,13 +1,0 @@
-export default function MakeBookButton() {
-    return (
-        <div
-            className="relative shadow-lg rounded flex justify-center items-center text-xs font-medium text-gray-700 bg-white ml-5 mb-10"
-            style={{
-                width: '7vw',
-                height: '14vh',
-            }}
-        >
-            +
-        </div>
-    );
-}
